# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Back2 Newsletters",
        "description": "Documentation Swagger pour la gestion des newsletters et archives.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Liste tous les articles",
                "operationId": "89181b74a8aa4d580750ede6aa32f92a",
                "responses": {
                    "200": {
                        "description": "Liste des articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Créer un nouvel article",
                "operationId": "ce06e2544b80bc76b184c771c0fb2e7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "typePart"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "video_url": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "typePart": {
                                        "type": "string",
                                        "enum": [
                                            "communique",
                                            "discours",
                                            "interview",
                                            "autre"
                                        ]
                                    },
                                    "publier_le": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "retirer_le": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/articles/{article}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Détails d'un article",
                "operationId": "eebdfd917b989bf9c6c551e98bc776ff",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Mettre à jour un article",
                "operationId": "b77d56981a88ec4be0c93e06d334f74f",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "typePart": {
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "video_url": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "publier_le": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "retirer_le": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Supprimer un article",
                "operationId": "9ee91bc5331371f5a1b884fd17b19279",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article supprimé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/articles/{article}/publish": {
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Publier un article",
                "operationId": "df63bd101fd094096d2440adafba2e99",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article publié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/articles/{article}/unpublish": {
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Retirer un article",
                "operationId": "79919e8cdfc6a083692afc8456415bb6",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article retiré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Article": {
                "title": "Article",
                "description": "Modèle représentant un article de blog ou une actualité",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Titre de l'article"
                    },
                    "content": {
                        "type": "string",
                        "example": "Contenu de l'article..."
                    },
                    "slug": {
                        "type": "string",
                        "example": "titre-de-l-article"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "images/photo.jpg",
                        "nullable": true
                    },
                    "video_url": {
                        "type": "string",
                        "example": "videos/clip.mp4",
                        "nullable": true
                    },
                    "full_image_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/images/photo.jpg",
                        "nullable": true
                    },
                    "full_video_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/videos/clip.mp4",
                        "nullable": true
                    },
                    "typePart": {
                        "type": "string",
                        "example": "communique"
                    },
                    "auteur": {
                        "description": "ID de l'auteur",
                        "type": "integer",
                        "example": 1
                    },
                    "publier_le": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "retirer_le": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}